pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "it-spend-dashboard"
version = "0.1.0"
description = "Analytics pipeline and HTML dashboard for IT department operating expenses from 1C CSV exports."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Codex" }
]
dependencies = [
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "Jinja2>=3.1.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
